fix(cli): isolate and retry voice authentication
